requirements.txt
streamlit>=1.40.0
python-dotenv>=1.0.0
google-adk>=1.0.0
litellm>=1.40.0
httpx>=0.27.0
ddgs>=9.0.0
beautifulsoup4>=4.12.0
